修改枚举格式，忽略字段和支持字段

+++ MJExtensionExample/main.m
@@ -77,7 +77,7 @@ void keyValues2object()
     MJUser *user = [MJUser mj_objectWithKeyValues:dict];
     
     // 3.打印MJUser模型的属性
-    MJExtensionLog(@"name=%@, icon=%@, age=%zd, height=%@, money=%@, sex=%d, gay=%d", user.name, user.icon, user.age, user.height, user.money, user.sex, user.gay);
+    MJExtensionLog(@"name=%@, icon=%@, age=%zd, height=%@, money=%@, sex=%ld, gay=%d", user.name, user.icon, user.age, user.height, user.money, (long)user.sex, user.gay);
 }
 
 /**
